requirements.txt
Django==1.10.4
django-registration-redux==1.4
djangorestframework==3.5.3
pytz==2016.10
